feat: adiciona documentação no swagger

+++ src/services/usuario.service.ts
@@ -1,47 +1,47 @@
-import { Usuario } from '../entities/usuario';
-import { UsuarioRepository } from '../repositories/usuario.repository';
-
-import { UsuarioCadastroDTO } from '../dto/usuarioDTO';
-
-import bcrypt from 'bcrypt';
+import { Usuario } from "../entities/usuario";
+import { UsuarioRepository } from "../repositories/usuario.repository";
+import { UsuarioCadastroDTO } from "../dto/usuarioDTO";
+import bcrypt from "bcrypt";
 
 export class UsuarioService {
-    
   private usuarioRepository = new UsuarioRepository();
 
   async cadastrarUsuario(dto: UsuarioCadastroDTO): Promise<Usuario> {
-
-    const existente = await this.usuarioRepository.findByEmail(dto.email)
-    if (existente) throw new Error('Email já cadastrado');
+    const emailExistente = await this.usuarioRepository.findByEmail(dto.email);
+    if (emailExistente) throw new Error("Email já cadastrado");
 
     const senhaHash = await bcrypt.hash(dto.senha, 10);
+
     const usuario = new Usuario(null, dto.nome, dto.email, senhaHash);
 
     return await this.usuarioRepository.create(usuario);
-
   }
 
   async listarUsuarios(): Promise<Usuario[]> {
-    
-    if(!this.usuarioRepository.findAll){
-      throw new Error('Erro ao buscar os usuários')
+    if (!this.usuarioRepository.findAll()) {
+      throw new Error("Erro ao buscar os usuários");
     }
-
     return await this.usuarioRepository.findAll();
-
   }
 
-  async atualizarUsuario(id: number, nome: string, email: string, senha?: string): Promise<Usuario> {
+  async buscarUsuario(nome: string): Promise<Usuario[]> {
 
+    const usuario = await this.usuarioRepository.findByName(nome);
+    if (!usuario) throw new Error("Usuário não encontrado");
+
+    return usuario;
+  }
+
+  
+  async atualizarUsuario(id: number, dto: UsuarioCadastroDTO): Promise<Usuario> {
     const usuario = await this.usuarioRepository.findById(id);
+    if (!usuario) throw new Error("Usuário não encontrado");
 
-    if (!usuario) throw new Error('Usuário não encontrado');
+    usuario.setNome(dto.nome);
+    usuario.setEmail(dto.email);
 
-    usuario.setNome(nome);
-    usuario.setEmail(email);
-
-    if (senha) {
-      const senhaHash = await bcrypt.hash(senha, 10);
+    if (dto.senha) {
+      const senhaHash = await bcrypt.hash(dto.senha, 10);
       usuario.setSenha(senhaHash);
     }
 
@@ -50,10 +50,9 @@ export class UsuarioService {
   }
 
   async excluirUsuario(id: number): Promise<void> {
+    const usuario = await this.usuarioRepository.findById(id);
+    if (!usuario) throw new Error("Usuário não encontrado");
 
     await this.usuarioRepository.delete(id);
-
   }
-
-
 }
